mremap.2: Add note about mremap() with locked areas
When mremap() is used to move or expand a mapping that is locked
with mlock() or equivalent it will attempt to populate the new
area.  However, like mmap(MAP_LOCKED), mremap() will not fail if
the area cannot be populated.  Also like mmap(MAP_LOCKED) this
might come as a surprise to users and should be noted.

Signed-off-by: Eric B Munson <emunson@akamai.com>
Acked-by: Michal Hocko <mhocko@suse.com>
Cc: Michal Hocko <mhocko@suse.cz>
Cc: David Rientjes <rientjes@google.com>
Cc: linux-man@vger.kernel.org
Cc: linux-mm@kvack.org
Cc: linux-kernel@vger.kernel.org
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>