termios.3: Add missing details on behaviour of PARMRK
For a serial terminal, with a specific configuration, input bytes
with value 0377 are passed to the program as two bytes, 0377 0377.

This (correct) behaviour is described in the documentation of the
GNU C Library
(https://www.gnu.org/software/libc/manual/html_node/Input-Modes.html#Input-Modes)
but not in the termios.3 man page.

Problematic configuration: INPCK set, IGNPAR not set, PARMRK set,
ISTRIP not set.

This man page problem affects several users. Examples:
* http://sourceforge.net/p/ftdi-usb-sio/mailman/message/4079724/
* http://mailman.uclinux.org/pipermail/uclinux-dev/2006-November/040984.html
* ...

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>