nsswitch.conf.5: Update NSS compatibility mode description
From the current description of NSS compatibility mode it seems
that /etc/passwd is the only file where special entries are
permitted.  But "compat" service can also be specified for group
and shadow databases, so this needs to be changed.

The list of special entries is for passwd database only, group
and shadow databases are not mentioned.  Because group database
does not support netgroup special entries and it deals with
groups, not users, it is better to make a separate list
of entries for it.

It is true that the default source for the compat pseudo-databases
is "nis", but it can be overridden by any NSS service, not just
"nisplus". Even "compat" itself can be specified as the source for
the pseudo-databases, but doing that of course leads to infinite
recursion, so it makes sense to disallow that.

The information was obtained from glibc source code, namely from
the following files:
nis/nss_compat/compat-pwd.c
nis/nss_compat/compat-grp.c
nis/nss_compat/compat-spwd.c

Signed-off-by: Nikola Forró <nforro@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>