wait.2, wait4.2: Rename the "status" argument to "wstatus"
The fact that exit(3)/_exit(2) has an argument called
"status" and the same name is used in the arguments to the
wait*() calls can a little too easily lead the user into
thinking that the two arguments hold the same information,
when of course they don't. So, use a different name
for the argument of the wait*() functions, to reduce
the chances of such confusion.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>