locale.5: Complete LC_COLLATE
Here's the first attempt to (almost) complete the locale.5 manual
page by documenting all (but perhaps one) of the missing
LC_COLLATE keywords.

I think the LC_COLLATE section is still not enough to be used as
the only source when writing collation rules from scratch but
perhaps that's not even needed, it could be also thought that
the section 5 pages merely describe the format used in the files.
Naturally more information could be added later on top of this
patch.

Few notes:

- AFAICS coll_weight_max is not used anywhere in glibc
- I'm not aware of any C library implementation on Linux (for
  which this manual page would be relevant) which would
  implement the POSIX options not supported by glibc
- the glibc specific script keyword could still be documented

Sources:

http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap07.html
http://www.open-std.org/jtc1/SC22/WG20/docs/n972-14652ft.pdf

PS. A couple of unrelated comment clean-ups slipped in as well,
sorry about those.