lirc.4: Update after upstreamed lirc.h, bugfixes.
The lirc.h header has landed in the kernel, and the kernel docs
has been updated all of which reflected in this patch.

Here is still an open issue with duplicated info in the kernel
docs and the manpage. Eventually, this should be addressed but
I frankly don't know how. In the meantime, acknowledge the fact
that the kernel docs is the ultimate source