recv.2: With pending 0-length datagram read() and recv() with flags == 0 differ
Quoting Tom:

The statement "With a zero flags argument, recv() is equivalent to
read(2)." is not correct. In the case of passing a UDP socket an
empty buffer the two calls differ. read(2) will be a noop (as it
rightly says in its manpage), but recv(2) will discard the
packet.

We ran into this in networkd, as we use FIONREAD to determine the
buffer size (and allocate the right buffer), so in case someone
passed us an empty packet we would end up in a busy loop when we
were using read(2).  Changing to recv(2) fixed the issue
[https://github.com/systemd/systemd/pull/3299].

Reported-by: Tom Gundersen <teg@jklm.no>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>