locale.5: Adjust conformance
Since we now know that glibc does not support all the keywords
mentioned in ISO/IEC TR 14652 [1] and in general glibc aims to
be conforming to POSIX first and foremost, I think it's best
just to drop the reference to ISO/IEC TR 14652 from the man page.

1) http://www.open-std.org/jtc1/SC22/WG20/docs/n972-14652ft.pdf

After this I think the locale related man pages are finally as
complete as they need to be.