rename.2: Relocate some text text
Relocate text noting that there may be a window when 'oldpath' and
'newpath' refer to the same file. Logically, this text appears to
belong near the text noting that an existing 'newpath' will be
atomically replaced.  (In ancient versions of the page, these two
pieces of text were closer together.)

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>