perf_event_open.2: Add cycles field in LBR records
Linux 4.3 (71ef3c6b9d4665ee7afbbe4c208a98917dcfc32f)
added a cycles field to the PERF_SAMPLE_BRANCH_STACK
last branch records.

The kernel commit was a bit vague on this, but you can find
a few more details on this in the Intel Architectural Manual
vol3B.  The field indicates the number of core cycles elapsed
since the previous update to the LBR stack.

This feature is only found on Skylake and newer Intel chips,
as well as Intel Atom Goldmont chips.  I'm not sure if it's
worth adding this info to the manpage, as it seems a bit
specific and will probably get rapidly out of date.

Signed-off-by: Vince Weaver <vincent.weaver@maine.edu>