perf_event_open.2: Update time_shift sample code
Linux 4.3 (b20112edeadf0b8a1416de061caa4beb11539902) improved
the accuracy of the clock/ns conversion routines.  As a result
the shift factor can now be 32.  This value is directly
exported in the perf_event_open() mmap page, and this
potentially breaks the sample code that shifts 1 left by
the shift value.

Add a cast in the sample code so that a proper 64-bit value
results from the shift.  This is the same change that was
made to the sample code in include/uapi/linux/perf_event.h
in Linux 4.4 (b9511cd761faafca7a1acc059e792c1399f9d7c6).

Signed-off-by: Vince Weaver <vincent.weaver@maine.edu>