getrandom.2, random.4: Consolidate and improve discussion on usage of randomness
Currently, recommendations on how to consume randomness are
spread across both getrandom(2) and random(4) and the general
opinion seems to be that the text in getrandom(2) does a
somewhat better job. Consolidate the discussion to a single
page (getrandom(2)) and address some of the concerns
expressed about the existing text in random(4).

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>