fuse.4: New page describing /dev/fuse
This is my writeup of a basic description of /dev/fuse after
playing with it for a few hours today. It is of course woefully
incomplete, and since I neither have a use case nor am working
on this code, I will not be in a position to expand it in the
near future. However, I'm hoping this could still serve as a
handy reference for others looking at this interface.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>