poll.2, select.2: Add a reference to the sigset discussion in sigprocmask(2)
A little while back, I added a note to sigprocmask.2 that
discussed the difference between the libc's and the kernel's
sigset_t structures.  I added that note, because I saw this being
done wrong in a tool tracing system calls (causing subtle bugs).
As it turns out, the same bugs existed for ppoll and pselect, for
the same reason. I'm hoping by adding the reference here, future
writers of similar tools will find that discussion and not make
the same mistake.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>