inode.7: New page with information about inodes
David Howells provided a statx(2) page that duplicated
much of the information from form stat(2). Such duplication
is undesirable, and there are two possible solutions:

* merge the statx() description into the existing
  stat(2) page.
* move the common information in stat(2) and statx(2)
  to a new page.

The first option suffers from the fact that the stat(2) page
is already rather overloaded with a description of four APIs;
adding a fifth would make the page even more unwieldy!
This patch opts for the second solution. After this,
we must remove the duplicated material from stat(2) and
statx(2).

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>