getcontext.3: Exemplar structure should use 'ucontext_t'.
In glibc we just finished a round of purging 'struct ucontext'
which is not in the POSIX reserved namespace of *_t tags. This has
some consequences to applications using the non-standard struct
ucontext:
https://sourceware.org/glibc/wiki/Release/2.26#Removal_of_.27struct_ucontext.27
but it also fixes a namespace conformance issue which is always a
longterm pain for large portable programs.

It was noted by Peter Maydell
(https://sourceware.org/bugzilla/show_bug.cgi?id=21457) that the
linux man pages still had references to 'struct ucontext' but only
in the form of an exemplar structure.

The patch fixes the exemplar to match what is in glibc
and therefore it won't ever suggest users can use 'struct
ucontext'.

Signed-off-by: Carlos O'Donell <carlos@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>