ld.so.8: Expand DT_RUNPATH details.
ld.so.8: Expand DT_RUNPATH details.

Every 3 years we get asked why DT_RUNPATH doesn't work like DT_RPATH.
The most recent question was here:
https://www.sourceware.org/ml/libc-help/2017-06/msg00013.html

We need to expand the description of DT_RUNPATH to cover this situation
and explain that the DT_RUNPATH entries apply only to the immediate
DT_NEEDED, not that of another, say dlopen'd child object.

Applies to master.

Signed-off-by: Carlos O'Donell <carlos@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>