cmsg.3: Add a scatter/gather buffer to sample code
Add a simple scatter/gather buffer to the cmsg sample code.

It appears that the scatter/gather buffer is required even though
it may not be used. In my testing (on Fedora 24
4.8.7-200.fc24.x86_64 as well as a 4.11 based kernel) return value
of sendmsg() was the number of bytes in the scatter/gather buffer
(1 with the patch below).

Without the iovec buffer, sendmsg() returns 0 which means no bytes
were transferred and the corresponding recvmsg() blocks
indefinitely.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>