tzfile.5: Sync from tzdb upstream
This makes tzfile.5 a copy of the tzdb version, except that the
tzdb version's first line is replaced by man-pages boilerplate.
The new version documents version 3 format, among other things.
Also, it removes the "Summary of the timezone information file
format" section, which should no longer be needed due to
improvements in the the part of the man page.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>