locale.5: Refer to existing locales for encoding details
Mike Fabian and Rafał Luzynski were recently named as glibc
localedata maintainers [1] and after that there's been active
development on this front, including discussion whether it would
be ok to use ASCII or some other encoding as values for actual
locate data.

Since I don't think it would make sense to try to have different
explanation for each glibc version on the locale(5) man page, I'm
proposing that we apply the below patch so that we refer to
existing locale definition files in general and not spell out the
exact format or any certain locale as a definitive guideline.

If the situation changes in the future or new a new convention
meant to last forever is created then perhaps Mike and Rafał can
provide an update then as needed.

1) https://sourceware.org/ml/libc-alpha/2017-07/msg00477.html
2) https://sourceware.org/ml/libc-alpha/2017-07/msg00807.html

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>