getrlimit.2: Mention unit used by RLIMIT_CORE and RLIMIT_FSIZE
It would have been obvious that these limits are in bytes, except that
"ulimit -a" in at least bash, dash and zsh says that they're in blocks.
This confused me, so I had to check the kernel source code.

My understanding is that they are indeed in bytes, so mention this
information in the man page.

Signed-off-by: Thiago Jung Bauermann <bauerman@linux.vnet.ibm.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>