membarrier.2: Update example to take TSO into account
The existing example given specifically states that it focus on
x86 (TSO memory model), but gives a read-read vs write-write
ordering example, even though this scenario does not require
explicit barriers on TSO.

So either we change the example architecture to a weakly-ordered
architecture, or we change the example to a scenario requiring
barriers on x86.

Let's stay on x86, but provide a Dekker as example instead.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
CC: Michael Kerrisk <mtk.manpages@gmail.com>
CC: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Link: https://stackoverflow.com/questions/45970525/is-the-example-in-the-membarrier-man-page-pointless-in-x86
Link: https://lwn.net/Articles/573436/
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>